requirements.txt
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.0.0
PyGLM==1.1.8
PyOpenGL==3.1.5
pyparsing==2.4.6
pyqtgraph==0.10.0
scipy==1.2.1
pyside2
